Add property-based & example tests for full boards, fix bug w/ pointers
Totally missed that I was checking pointer equivalence—and not the
actual value of each space.

This wasn't detected previously, because the other tests would
directly assign the reference to the player token. However, in the full
board tests, I just so happened to create a COPY of the player's token
before assignment, which led to different pointers in every space of the
board. With the old implementation, this resulted in it being impossible
to win because no pointers would ever point to the same address. The new
implementation will dereference the pointer to check its actual value,
preventing this unintended side effect of using different references to
the same token.