Heavily refactor Board to track player turns
And add invariants for assigning spaces after game has ended.

This was necessary because:

- minimax needs to know which players are on the board, and I didn't
want to give minimax more knowledge about the game

This made things simpler because there's no need to track which player
gets assigned to which space, or whose turn it is outside of the board.
Now Game simply gets the currently active player (tracked by turns on
the board) and asks them to choose.

The Board API change also led to DRYing up a lot of logic in tests for
"filling" the board with spaces, resulting in some utility functions for
interleaving lists of ints and such to make tests slightly more
readable.